+++ IsekaiMod/Content/Features/IsekaiProtagonist/OverpoweredAbility/AuraOfRighteousWrath.cs
@@ -17,16 +17,16 @@ using Kingmaker.UnitLogic.FactLogic;
 using Kingmaker.UnitLogic.Mechanics.Components;
 using Kingmaker.Utility;
 
-namespace IsekaiMod.Content.Features.IsekaiProtagonist.SpecialPower
+namespace IsekaiMod.Content.Features.IsekaiProtagonist.OverpoweredAbility
 {
-    class AuraOfZealousWrath
+    class AuraOfRighteousWrath
     {
         public static void Add()
         {
-            var Icon_AuraOfZealousWrath = AssetLoader.LoadInternal("Features", "ICON_AURA_ZEALOUS_WRATH.png");
-            var AuraOfZealousWrathEnchantment = Helpers.CreateWeaponEnchantment("AuraOfZealousWrathEnchantment", bp => {
-                bp.m_EnchantName = Helpers.CreateString("AuraOfZealousWrathEnchantment.Name", "Aura of Zealous Wrath");
-                bp.m_Description = Helpers.CreateString("AuraOfZealousWrathEnchantment.Description", "This creature has two extra attacks and deal an additional 5d6 physical damage. "
+            var Icon_AuraOfRighteousWrath = AssetLoader.LoadInternal("Features", "ICON_AURA_RIGHTEOUS_WRATH.png");
+            var AuraOfRighteousWrathEnchantment = Helpers.CreateWeaponEnchantment("AuraOfRighteousWrathEnchantment", bp => {
+                bp.m_EnchantName = Helpers.CreateString("AuraOfRighteousWrathEnchantment.Name", "Overpowered Ability — Aura of Righteous Wrath");
+                bp.m_Description = Helpers.CreateString("AuraOfRighteousWrathEnchantment.Description", "This creature has two extra attacks and deal an additional 5d6 physical damage. "
                     + "It also has additional sneak attack damage.");
                 bp.AddComponent<WeaponConditionalDamageDice>(c => {
                     c.Damage = new DamageDescription()
@@ -49,11 +49,11 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.SpecialPower
                     c.Conditions = ActionFlow.EmptyCondition();
                 });
             });
-            var AuraOfZealousWrathBuff = Helpers.CreateBuff("AuraOfZealousWrathBuff", bp => {
-                bp.SetName("Aura of Zealous Wrath");
+            var AuraOfRighteousWrathBuff = Helpers.CreateBuff("AuraOfRighteousWrathBuff", bp => {
+                bp.SetName("Overpowered Ability — Aura of Righteous Wrath");
                 bp.SetDescription("This creature has two extra attacks and deal an additional 5d6 physical damage. It also has additional sneak attack damage.");
                 bp.IsClassFeature = true;
-                bp.m_Icon = Icon_AuraOfZealousWrath;
+                bp.m_Icon = Icon_AuraOfRighteousWrath;
                 bp.AddComponent<BuffExtraAttack>(c => {
                     c.Number = 2;
                 });
@@ -67,53 +67,55 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.SpecialPower
                     c.m_Progression = ContextRankProgression.Div2;
                 });
                 bp.AddComponent<BuffEnchantAnyWeapon>(c => {
-                    c.m_EnchantmentBlueprint = AuraOfZealousWrathEnchantment.ToReference<BlueprintItemEnchantmentReference>();
+                    c.m_EnchantmentBlueprint = AuraOfRighteousWrathEnchantment.ToReference<BlueprintItemEnchantmentReference>();
                     c.Slot = EquipSlotBase.SlotType.PrimaryHand;
                 });
                 bp.AddComponent<BuffEnchantAnyWeapon>(c => {
-                    c.m_EnchantmentBlueprint = AuraOfZealousWrathEnchantment.ToReference<BlueprintItemEnchantmentReference>();
+                    c.m_EnchantmentBlueprint = AuraOfRighteousWrathEnchantment.ToReference<BlueprintItemEnchantmentReference>();
                     c.Slot = EquipSlotBase.SlotType.SecondaryHand;
                 });
                 bp.AddComponent<BuffEnchantAnyWeapon>(c => {
-                    c.m_EnchantmentBlueprint = AuraOfZealousWrathEnchantment.ToReference<BlueprintItemEnchantmentReference>();
+                    c.m_EnchantmentBlueprint = AuraOfRighteousWrathEnchantment.ToReference<BlueprintItemEnchantmentReference>();
                     c.Slot = EquipSlotBase.SlotType.AdditionalLimb;
                 });
             });
-            var AuraOfZealousWrathArea = Helpers.CreateBlueprint<BlueprintAbilityAreaEffect>("AuraOfZealousWrathArea", bp => {
+            var AuraOfRighteousWrathArea = Helpers.CreateBlueprint<BlueprintAbilityAreaEffect>("AuraOfRighteousWrathArea", bp => {
                 bp.m_TargetType = BlueprintAbilityAreaEffect.TargetType.Ally;
                 bp.Shape = AreaEffectShape.Cylinder;
-                bp.Size = new Feet(40);
+                bp.Size = new Feet(120);
                 bp.Fx = new PrefabLink();
-                bp.AddComponent(AuraUtils.CreateUnconditionalAuraEffect(AuraOfZealousWrathBuff.ToReference<BlueprintBuffReference>()));
+                bp.AddComponent(AuraUtils.CreateUnconditionalAuraEffect(AuraOfRighteousWrathBuff.ToReference<BlueprintBuffReference>()));
             });
-            var AuraOfZealousWrathAreaBuff = Helpers.CreateBuff("AuraOfZealousWrathAreaBuff", bp => {
-                bp.SetName("Aura of Zealous Wrath");
-                bp.SetDescription("Allies within 40 feet of you has two extra attacks and deal an additional 5d6 physical damage. "
+            var AuraOfRighteousWrathAreaBuff = Helpers.CreateBuff("AuraOfRighteousWrathAreaBuff", bp => {
+                bp.SetName("Overpowered Ability — Aura of Righteous Wrath");
+                bp.SetDescription("Allies within 120 feet of you has two extra attacks and deal an additional 5d6 physical damage. "
                     + "They also gain a number of sneak attack dice equal to 1/2 your character level.");
-                bp.m_Icon = Icon_AuraOfZealousWrath;
+                bp.m_Icon = Icon_AuraOfRighteousWrath;
                 bp.IsClassFeature = true;
                 bp.m_Flags = BlueprintBuff.Flags.HiddenInUi;
                 bp.AddComponent<AddAreaEffect>(c => {
-                    c.m_AreaEffect = AuraOfZealousWrathArea.ToReference<BlueprintAbilityAreaEffectReference>();
+                    c.m_AreaEffect = AuraOfRighteousWrathArea.ToReference<BlueprintAbilityAreaEffectReference>();
                 });
             });
-            var AuraOfZealousWrathAbility = Helpers.CreateActivatableAbility("AuraOfZealousWrathAbility", bp => {
-                bp.SetName("Aura of Zealous Wrath");
-                bp.SetDescription("Allies within 40 feet of you has two extra attacks and deal an additional 5d6 physical damage. "
+            var AuraOfRighteousWrathAbility = Helpers.CreateActivatableAbility("AuraOfRighteousWrathAbility", bp => {
+                bp.SetName("Overpowered Ability — Aura of Righteous Wrath");
+                bp.SetDescription("Allies within 120 feet of you has two extra attacks and deal an additional 5d6 physical damage. "
                     + "They also gain a number of sneak attack dice equal to 1/2 your character level.");
-                bp.m_Icon = Icon_AuraOfZealousWrath;
-                bp.m_Buff = AuraOfZealousWrathAreaBuff.ToReference<BlueprintBuffReference>();
+                bp.m_Icon = Icon_AuraOfRighteousWrath;
+                bp.m_Buff = AuraOfRighteousWrathAreaBuff.ToReference<BlueprintBuffReference>();
                 bp.DoNotTurnOffOnRest = true;
             });
-            var AuraOfZealousWrathFeature = Helpers.CreateFeature("AuraOfZealousWrathFeature", bp => {
-                bp.SetName("Aura of Zealous Wrath");
-                bp.SetDescription("Allies within 40 feet of you has two extra attacks and deal an additional 5d6 physical damage. "
+            var AuraOfRighteousWrathFeature = Helpers.CreateFeature("AuraOfRighteousWrathFeature", bp => {
+                bp.SetName("Overpowered Ability — Aura of Righteous Wrath");
+                bp.SetDescription("Allies within 120 feet of you has two extra attacks and deal an additional 5d6 physical damage. "
                     + "They also gain a number of sneak attack dice equal to 1/2 your character level.");
-                bp.m_Icon = Icon_AuraOfZealousWrath;
+                bp.m_Icon = Icon_AuraOfRighteousWrath;
                 bp.AddComponent<AddFacts>(c => {
-                    c.m_Facts = new BlueprintUnitFactReference[] { AuraOfZealousWrathAbility.ToReference<BlueprintUnitFactReference>() };
+                    c.m_Facts = new BlueprintUnitFactReference[] { AuraOfRighteousWrathAbility.ToReference<BlueprintUnitFactReference>() };
                 });
             });
+
+            OverpoweredAbilitySelection.AddToSelection(AuraOfRighteousWrathFeature);
         }
     }
 }