+++ CHANGELOG.md
@@ -5,10 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added `Aura of Righteous Wrath` Overpowered ability.
 ### Changed
 - Patched Prestige classes to allow `Isekai Protagonist` spellbook to be selected.
 - Changed `Isekai Protagonist` to count as both an arcane and divine caster.
 - Changed `Supersonic Combat` and `Graceful Combat` to affect damage for ranged attacks.
+- Renamed `Infinite Inventory` to `Infinite Space`.
 
 ## [3.0.1] - 2022-12-12
 ### Changed